requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
transformers>=4.44.0
prometheus-fastapi-instrumentator
loguru
gradio
boto3
python-dotenv